Update on "Add MDETR image backbone"
This commit adds support for the ResNet-101 image backbone from MDETR.
This is basically just a wrapper around TorchVision's ResNet class
with support for fetching intermediate layers from the model so that the final pooling and
FC layers are not used. We also add the nested tensor and frozen batch norm classes from
the MDETR repo, ultimately we may factor out the nested tensor class later on.

Also added a test to confirm the results match those obtained by using MDETR's image backbone directly.

This is an initial version, will update docstrings etc if the high-level implementation looks good.

Tested with
```
$ python -m pytest -v test/modules/encoders/test_mdetr_image_encoder.py
test/modules/encoders/test_mdetr_image_encoder.py::TestMDETRImageEncoder::test_resnet_101_forward PASSED

[ghstack-poisoned]
